LayoutTests/fast/mediastream/error-expected.txt

Tests UserMedia error callback using the mock service.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.code is error.PERMISSION_DENIED
PASS successfullyParsed is true

TEST COMPLETE
